requirements.txt
flask
flask-cors
langchain
langchain-community
langchain-google-genai
langchain-core
langchain-chroma
python-dotenv
chromadb
docx2txt
pandas
openpyxl
pywin32
pypdf 